+++ packages/test-snaps/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.16.0]
+
+### Added
+
+- Add `onSettingsPage` example to preinstalled example Snap ([#2911](https://github.com/MetaMask/snaps/pull/2911), [#2962](https://github.com/MetaMask/snaps/pull/2962))
+
 ## [2.15.4]
 
 ### Changed
@@ -241,7 +247,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fix NPM package name of the network access snap ([#1621](https://github.com/MetaMask/snaps/pull/1621))
 
-[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/test-snaps@2.15.4...HEAD
+[Unreleased]: https://github.com/MetaMask/snaps/compare/@metamask/test-snaps@2.16.0...HEAD
+[2.16.0]: https://github.com/MetaMask/snaps/compare/@metamask/test-snaps@2.15.4...@metamask/test-snaps@2.16.0
 [2.15.4]: https://github.com/MetaMask/snaps/compare/@metamask/test-snaps@2.15.3...@metamask/test-snaps@2.15.4
 [2.15.3]: https://github.com/MetaMask/snaps/compare/@metamask/test-snaps@2.15.2...@metamask/test-snaps@2.15.3
 [2.15.2]: https://github.com/MetaMask/snaps/compare/@metamask/test-snaps@2.15.1...@metamask/test-snaps@2.15.2